+++ .vscode/settings.json
@@ -8,6 +8,7 @@
     "mobin",
     "Mori",
     "nextui",
+    "nuwa",
     "Reza",
     "Workmark"
   ],
