+++ package.json
@@ -82,6 +82,8 @@
     "eslint": "^8.57.0",
     "eslint-config-next": "14.0.3",
     "postcss": "^8.4.38",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.5",
     "prisma": "^5.13.0",
     "tailwindcss": "^3.4.3",
     "ts-node": "^10.9.2",
