+++ src/app/(web)/posts/[slug]/page.tsx
@@ -1,118 +1,24 @@
-import { Separator } from "@/components/ui/separator";
-import { markdownToHTML } from "@/lib/markdownToHTML";
-
-import { getPostBySlug } from "@/service/posts.service";
-
-import Image from "next/image";
-
-import { RxDotFilled } from "react-icons/rx";
-
-import placeholderImage from "public/images/placeholder.png";
 import BlurImage from "@/components/global/image-blur";
 import ClapContainer from "@/components/posts/claps/clap-container";
+import { Separator } from "@/components/ui/separator";
+import { markdownToHTML } from "@/lib/markdownToHTML";
 import { convertDateToDayMonthAndYear } from "@/lib/utils";
-import { PostModel } from "@/types/post.type";
-import { Metadata, ResolvingMetadata } from "next";
+import { generateSchemaJson } from "@/lib/utils/jsonLD/postsJsonLd";
+import { generateMetadataForPost } from "@/lib/utils/metadata/postsMetadata";
+import { getPostBySlug } from "@/service/posts.service";
+import { Metadata } from "next";
+import Image from "next/image";
 import { notFound } from "next/navigation";
+import placeholderImage from "public/images/placeholder.png";
+import { RxDotFilled } from "react-icons/rx";
 import PostNavigationGroup from "../nav";
-import { getHomepageOGImageLink } from "@/lib/utils/imageUtils";
 
-export async function generateMetadata(
-  { params }: Props,
-  parent: ResolvingMetadata
-): Promise<Metadata> {
+export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const slug = params.slug;
   const data = await getPostBySlug(slug);
   if (!data) notFound();
 
-  const metadata: Metadata = {
-    robots: {
-      index: true,
-      follow: true,
-      nocache: true,
-      googleBot: {
-        index: true,
-        follow: true,
-        "max-snippet": -1,
-        noimageindex: false,
-        "max-video-preview": -1,
-        "max-image-preview": "large",
-      },
-    },
-    title: data.title,
-    openGraph: {
-      type: "article",
-      locale: "en_US",
-      siteName: "Mora Blog",
-      authors: data.author.name,
-      description: data.summary,
-      title: `${data.title} | Mora Blog`,
-      url: `https://www.mora-ed.com/posts/${slug}`,
-      publishedTime: data?.publishedAt?.toUTCString(),
-      modifiedTime: data?.updatedAt?.toUTCString(),
-      images: [
-        {
-          width: "1080",
-          height: "1080",
-          alt: data.title,
-          type: "image/jpeg",
-          url: data.mainImageUrl ?? getHomepageOGImageLink(),
-        },
-      ],
-    },
-    twitter: {
-      site: "@",
-      description: data.summary,
-      card: "summary_large_image",
-      images: [
-        {
-          width: "1080",
-          height: "1080",
-          alt: data.title,
-          type: "image/jpeg",
-          url: data.mainImageUrl ?? getHomepageOGImageLink(),
-        },
-      ],
-      siteId: "https://www.mora-ed.com",
-      title: `${data.title} | Mora Blog`,
-    },
-  };
-
-  return metadata;
-}
-
-function generateSchemaJson(post: PostModel) {
-  const jsonLd = {
-    name: post.title,
-    identifier: post.id,
-    headline: post.title,
-    "@type": "NewsArticle",
-    description: post.summary,
-    image: [post.mainImageUrl],
-    creator: [post.author.name],
-    dateCreated: post.createdAt,
-    dateModified: post.updatedAt,
-    "@context": "https://schema.org",
-    url: "https://www.mora-ed.com/posts/" + post.slug,
-    datePublished: post.publishedAt ?? post.createdAt,
-    mainEntityOfPage: "https://mora-ed.com/posts/" + post.slug,
-
-    author: {
-      "@type": "Person",
-      name: post.author.name,
-    },
-    publisher: {
-      name: "Mora Blog",
-      "@type": "Organization",
-      url: "https://www.mora-ed.com",
-      logo: {
-        "@type": "ImageObject",
-        url: "https://d1ntfq67otjmwh.cloudfront.net/mora-blog-files/1717948463820-logo-small.svg",
-      },
-    },
-  };
-
-  return jsonLd;
+  return generateMetadataForPost(data);
 }
 
 type Props = {
@@ -129,57 +35,57 @@ async function PostPage({ params }: Props) {
   const htmlContent = await markdownToHTML(data.body ?? "");
 
   return (
-    <div className='pt-8 flex flex-col gap-4 px-4 mb-8 max-w-screen-md mx-auto'>
+    <div className="mx-auto mb-8 flex max-w-screen-md flex-col gap-4 px-4 pt-8">
       {/* metadata */}
       <script
-        type='application/ld+json'
+        type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(schema) }}
       />
       <title>{data.title}</title>
 
       <PostNavigationGroup postId={data.id} />
 
-      <h1 className='text-3xl font-extrabold leading-9 pt-2'>{data.title}</h1>
+      <h1 className="pt-2 text-3xl font-extrabold leading-9">{data.title}</h1>
       <BlurImage
         width={"288"}
         height={"160"}
-        loading='lazy'
+        loading="lazy"
         alt={data.title + " " + "main image"}
         src={data.mainImageUrl ?? placeholderImage}
-        className='rounded-lg self-center w-full object-contain'
+        className="w-full self-center rounded-lg object-contain"
       />
-      <h3 className='text-lg leading-6 text-neutral-500 dark:text-neutral-400'>
+      <h3 className="text-lg leading-6 text-neutral-500 dark:text-neutral-400">
         {data.summary}
       </h3>
 
-      <div className='flex flex-wrap gap-2 justify-between'>
-        <section aria-labelledby='author' className='flex items-center gap-2'>
+      <div className="flex flex-wrap justify-between gap-2">
+        <section aria-labelledby="author" className="flex items-center gap-2">
           <Image
             width={32}
             height={32}
-            alt='Author Avatar'
-            className='rounded-full aspect-square object-cover'
+            alt="Author Avatar"
+            className="aspect-square rounded-full object-cover"
             src={data.author.avatarUrl ?? "/images/main-image-placeholder.png"}
           />
-          <span className='text-neutral-500 text-sm leading-tight'>
+          <span className="text-sm leading-tight text-neutral-500">
             {data.author.name}
           </span>
         </section>
 
-        <section className='flex gap-1 items-center text-neutral-500 text-sm'>
+        <section className="flex items-center gap-1 text-sm text-neutral-500">
           <time>
             {convertDateToDayMonthAndYear(data.publishedAt ?? data.createdAt)}
           </time>
           <RxDotFilled />
           <span>{data.readingTime} min read</span>
 
-          <ClapContainer slug={slug} className='pb-1' total={data.claps} />
+          <ClapContainer slug={slug} className="pb-1" total={data.claps} />
         </section>
       </div>
 
       <Separator />
 
-      <article className='prose dark:prose-invert prose-pre:bg-gray-100 dark:prose-pre:bg-gray-900 prose-pre:font-normal dark:prose-pre:text-white prose-img:rounded-xl prose-a:text-blue-600 max-w-full'>
+      <article className="prose max-w-full dark:prose-invert prose-a:text-blue-600 prose-pre:bg-gray-100 prose-pre:font-normal prose-img:rounded-xl dark:prose-pre:bg-gray-900 dark:prose-pre:text-white">
         <div dangerouslySetInnerHTML={{ __html: htmlContent }} />
       </article>
     </div>
